Mistake on naming activities made numbers not hit
While merging the names of activities, I initially merged the activity ID column with the activity table, apart from the rest of the main data set. This caused the activities to be shuffled when I brought the column back, and the results for the variables didn't correspond to the right activities. On this version, I now merge the whole data set with the activity table, solving the issue.